consolidate work package mailing
