+++ app/views/work_package_mailer/_work_package_details.html.erb
@@ -1,0 +1,53 @@
+<%#-- copyright
+OpenProject is an open source project management software.
+Copyright (C) 2012-2021 the OpenProject GmbH
+
+This program is free software; you can redistribute it and/or
+modify it under the terms of the GNU General Public License version 3.
+
+OpenProject is a fork of ChiliProject, which is a fork of Redmine. The copyright follows:
+Copyright (C) 2006-2013 Jean-Philippe Lang
+Copyright (C) 2010-2013 the ChiliProject Team
+
+This program is free software; you can redistribute it and/or
+modify it under the terms of the GNU General Public License
+as published by the Free Software Foundation; either version 2
+of the License, or (at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program; if not, write to the Free Software
+Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+
+See COPYRIGHT and LICENSE files for more details.
+
+++#%>
+
+<h1><%= link_to work_package.to_s, work_package_url(work_package) %></h1>
+
+<ul>
+  <li><%= WorkPackage.human_attribute_name(:author)        %>: <%= work_package.author        %></li>
+  <li><%= WorkPackage.human_attribute_name(:status)        %>: <%= work_package.status        %></li>
+  <li><%= WorkPackage.human_attribute_name(:priority)      %>: <%= work_package.priority      %></li>
+  <li><%= WorkPackage.human_attribute_name(:assigned_to)   %>: <%= work_package.assigned_to   %></li>
+  <li><%= WorkPackage.human_attribute_name(:responsible)   %>: <%= work_package.responsible   %></li>
+  <li><%= WorkPackage.human_attribute_name(:category)      %>: <%= work_package.category      %></li>
+  <li><%= WorkPackage.human_attribute_name(:version) %>: <%= work_package.version %></li>
+  <li><%= WorkPackage.human_attribute_name(:start_date)    %>: <%= work_package.start_date    %></li>
+  <li><%= WorkPackage.human_attribute_name(:due_date)      %>: <%= work_package.due_date      %></li>
+  <li><%= WorkPackage.human_attribute_name(:estimated_hours)%>: <%= work_package.estimated_hours %></li>
+
+  <% if Setting.work_package_done_ratio != 'disabled' %>
+    <li><%= WorkPackage.human_attribute_name(:done_ratio)    %>: <%= work_package.done_ratio    %></li>
+  <% end %>
+
+  <% work_package.custom_field_values.each do |value| %>
+    <li><%= value.custom_field.name %>: <%= show_value(value) %></li>
+  <% end %>
+</ul>
+
+<%= format_text(work_package.description, attribute: :description, only_path: false, object: work_package, project: work_package.project) %>
