+++ app/views/work_package_mailer/_work_package_details.text.erb
@@ -27,24 +27,24 @@ See COPYRIGHT and LICENSE files for more details.
 
 ++#%>
 
-<%= issue.to_s %>
-<%= work_package_url(issue) %>
+<%= work_package.to_s %>
+<%= work_package_url(work_package) %>
 
-<%= WorkPackage.human_attribute_name(:author)        %>: <%= issue.author        %>
-<%= WorkPackage.human_attribute_name(:status)        %>: <%= issue.status        %>
-<%= WorkPackage.human_attribute_name(:priority)      %>: <%= issue.priority      %>
-<%= WorkPackage.human_attribute_name(:assigned_to)   %>: <%= issue.assigned_to   %>
-<%= WorkPackage.human_attribute_name(:responsible)   %>: <%= issue.responsible   %>
-<%= WorkPackage.human_attribute_name(:category)      %>: <%= issue.category      %>
-<%= WorkPackage.human_attribute_name(:version) %>: <%= issue.version %>
-<%= WorkPackage.human_attribute_name(:start_date)    %>: <%= issue.start_date    %>
-<%= WorkPackage.human_attribute_name(:due_date)      %>: <%= issue.due_date      %>
-<%= WorkPackage.human_attribute_name(:estimated_hours)%>: <%= issue.estimated_hours %>
+<%= WorkPackage.human_attribute_name(:author)        %>: <%= work_package.author        %>
+<%= WorkPackage.human_attribute_name(:status)        %>: <%= work_package.status        %>
+<%= WorkPackage.human_attribute_name(:priority)      %>: <%= work_package.priority      %>
+<%= WorkPackage.human_attribute_name(:assigned_to)   %>: <%= work_package.assigned_to   %>
+<%= WorkPackage.human_attribute_name(:responsible)   %>: <%= work_package.responsible   %>
+<%= WorkPackage.human_attribute_name(:category)      %>: <%= work_package.category      %>
+<%= WorkPackage.human_attribute_name(:version) %>: <%= work_package.version %>
+<%= WorkPackage.human_attribute_name(:start_date)    %>: <%= work_package.start_date    %>
+<%= WorkPackage.human_attribute_name(:due_date)      %>: <%= work_package.due_date      %>
+<%= WorkPackage.human_attribute_name(:estimated_hours)%>: <%= work_package.estimated_hours %>
 <% if Setting.work_package_done_ratio != 'disabled' %>
-<%=  WorkPackage.human_attribute_name(:done_ratio) %>: <%= issue.done_ratio %>
+<%=  WorkPackage.human_attribute_name(:done_ratio) %>: <%= work_package.done_ratio %>
 <% end %>
-<% issue.custom_field_values.each do |value| %>
+<% work_package.custom_field_values.each do |value| %>
 <%=  value.custom_field.name %>: <%= show_value(value) %>
 <% end %>
 
-<%= issue.description %>
+<%= work_package.description %>
