+++ app/views/work_package_mailer/watcher_changed.text.erb
@@ -26,12 +26,9 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 See COPYRIGHT and LICENSE files for more details.
 
 ++#%>
-<%= t("text_work_package_watcher_#{@action}", id: "##{@issue.id}", watcher_changer: @watcher_changer) %>
-<hr />
-<%= render partial: 'issue_details', locals: { issue: @issue } %>
-<p>
-  <%= format_text(t(:text_latest_note, note: last_work_package_note(@issue)),
-                    only_path: false,
-                    object: @issue,
-                    project: @issue.project) %>
-</p>
+
+<%= t("text_work_package_watcher_#{@action}", id: "##{@work_package.id}", watcher_changer: @watcher_changer) %>
+
+----------------------------------------
+<%= render partial: 'work_package_details', locals: { work_package: @work_package } %>
+<%= t(:text_latest_note, note: last_work_package_note(@work_package)) %>
