+++ spec/lib/open_project/hook_spec.rb
@@ -255,7 +255,7 @@ describe OpenProject::Hook do
       end
     end
     let!(:comparison_mail) do
-      UserMailer.work_package_watcher_changed(work_package, user, author, :added).deliver_now
+      WorkPackageMailer.watcher_changed(work_package, user, author, :added).deliver_now
       ActionMailer::Base.deliveries.last
     end
 
@@ -263,7 +263,7 @@ describe OpenProject::Hook do
       test_hook_controller_class.new.call_hook(:view_layouts_base_html_head)
 
       ActionMailer::Base.deliveries.clear
-      UserMailer.work_package_watcher_changed(work_package, user, author, :added).deliver_now
+      WorkPackageMailer.watcher_changed(work_package, user, author, :added).deliver_now
       mail2 = ActionMailer::Base.deliveries.last
 
       assert_equal comparison_mail.text_part.body.encoded, mail2.text_part.body.encoded
